Cleanup webpack config
