+++ lib/webpack-helpers/create-webpack-config.js
@@ -1,121 +1,15 @@
 'use strict';
-const fs = require('fs');
 const path = require('path');
-const webpack = require('webpack');
 const findupSync = require('findup-sync');
-const AssetsPlugin = require('assets-webpack-plugin');
 const resolvePkg = require('resolve-pkg');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const autoprefixer = require('autoprefixer');
 
-const { WEBPACK_ASSETS_BASENAME } = require('../constants');
-const ExcludeChunksPlugin = require('./exclude-chunks-plugin');
+const compact = require('../utils/compact-array');
 const safeWebpackUglifier = require('./safe-webpack-uglifier');
-const ModuleNotFoundPlugin = require('./module-not-found-plugin');
-const { generateTemplate } = require('./helpers');
+const webpackPlugins = require('./webpack-plugins');
+const webpackLoaders = require('./webpack-loaders');
 
-module.exports = createWebpackConfig;
-
-const compact = array => array.filter(Boolean);
-
-const getWebpackPlugins = urc => {
-  const publicAssetsPath = urc.publicAssetsPath.replace(/^\//, '');
-  return compact([
-    new ModuleNotFoundPlugin(urc.rootDirectory),
-    new AssetsPlugin({
-      path: path.resolve(urc.outputDirectory),
-      filename: WEBPACK_ASSETS_BASENAME,
-      processOutput: x => JSON.stringify(x, null, 2)
-    }),
-    new webpack.EnvironmentPlugin(urc.getClientEnvVars()),
-    new MiniCssExtractPlugin({
-      filename: `${publicAssetsPath}/css/[name]-[contenthash:8].css`,
-      chunkFilename: `${publicAssetsPath}/css/[name]-[contenthash:8].css`
-    }),
-
-    new HtmlWebpackPlugin({
-      inject: true,
-      excludeChunks: ['polyfill', 'runtime'],
-      minify: !urc.production
-        ? {}
-        : {
-            removeComments: true,
-            collapseWhitespace: true,
-            removeRedundantAttributes: true,
-            useShortDoctype: true,
-            removeEmptyAttributes: true,
-            removeStyleLinkTypeAttributes: true,
-            keepClosingSlash: true,
-            minifyJS: true,
-            minifyCSS: true,
-            minifyURLs: true
-          },
-      templateContent: ({ compilation, htmlWebpackPlugin }) =>
-        generateTemplate({
-          urc,
-          webpackCompilation: compilation,
-          publicPath: htmlWebpackPlugin.files.publicPath
-        })
-    }),
-    urc.hot && new webpack.HotModuleReplacementPlugin(),
-    ...urc.webpackPlugins,
-    // This should be the last plugin, since it removes certain
-    // chunks from webpack compilation.
-    // We are removing `runtime` chunk since it has already been
-    // inlined inside `HtmlWebpackPlugin`, hence writing it to
-    // disk would be wasteful.
-    new ExcludeChunksPlugin({
-      excludeChunks: ['runtime']
-    })
-  ]);
-};
-
-function createWebpackConfig(urc) {
+const webpackEntry = urc => {
   const jsEntryDir = path.dirname(urc.jsEntry);
-  const babelConfig = createBabelLoader(urc);
-  const publicAssetsPath = urc.publicAssetsPath.replace(/^\//, '');
-
-  const jsLoader = {
-    test: /\.jsx?$/,
-    exclude: [/[/\\\\]node_modules[/\\\\]/],
-    use: [babelConfig]
-  };
-  const fileLoader = {
-    loader: require.resolve('file-loader'),
-    exclude: [/\.(js|jsx)$/, /\.template$/, /\.json$/, /\.ejs$/],
-    options: {
-      name: `${publicAssetsPath}/media/[name].[hash:8].[ext]`
-    }
-  };
-  const cssLoader = {
-    test: /\.css$/,
-    use: [
-      urc.production
-        ? MiniCssExtractPlugin.loader
-        : require.resolve('style-loader'),
-      { loader: require.resolve('css-loader'), options: { importLoaders: 1 } },
-      {
-        loader: require.resolve('postcss-loader'),
-        options: {
-          ident: 'postcss',
-          plugins: [
-            autoprefixer({
-              // TOFIX if a user is supplying browserslist via urc,
-              // do we set an env var for everyone to read ?
-              browsers: urc.browserslist && urc.browserslist[urc.mode]
-            }),
-            ...urc.postcssPlugins
-          ]
-        }
-      }
-    ]
-  };
-
-  const sourceNodeModules = findupSync('node_modules', {
-    cwd: jsEntryDir
-  });
-
   const entry = {
     main: compact([
       urc.liveReload && require.resolve('webpack-dev-server/client') + '?/', // TOFIX wonder if we need to add localhost or not?
@@ -138,47 +32,34 @@ function createWebpackConfig(urc) {
   if (urc.polyfill === false) {
     delete entry.polyfill;
   }
+  return entry;
+};
+
+module.exports = function createWebpackConfig(urc) {
+  const publicAssetsPath = urc.publicAssetsPath.replace(/^\//, '');
+  const sourceNodeModules = findupSync('node_modules', {
+    cwd: path.dirname(urc.jsEntry)
+  });
+
   const config = {
+    bail: urc.production,
+    devtool: urc.production ? 'source-map' : 'cheap-module-source-map',
+    entry: webpackEntry(urc),
     mode: urc.production ? 'production' : 'development',
-    entry,
-    output: {
-      path: path.join(urc.outputDirectory),
-      publicPath: path.join(urc.siteBasePath, '/'), // TOFIX clean up this patching (/) of siteBasePath and also publicAssetsApth
-      filename: urc.production
-        ? `${publicAssetsPath}/js/[name]-[chunkhash:10].js`
-        : `${publicAssetsPath}/js/[name].js`,
-      chunkFilename: urc.production
-        ? `${publicAssetsPath}/js/[name]-[chunkhash:10].chunk.js`
-        : `${publicAssetsPath}/js/[name].chunk.js`
-    },
-    resolve: {
-      alias: {
-        // This is added to make Underreact symlinking work with npm <=3,
-        '@babel/runtime': path.dirname(
-          require.resolve('@babel/runtime/package.json')
-        )
-      }
-    },
     module: {
       rules: [
         {
-          oneOf: compact([
-            jsLoader,
-            cssLoader,
-            ...urc.webpackLoaders,
-            fileLoader
-          ])
+          oneOf: webpackLoaders(urc)
         }
       ]
     },
-    plugins: getWebpackPlugins(urc),
-    // Loader names need to be strings, and to allow them to be looked
-    // up within this module's nested dependencies, not just the user's
-    // node_modules, we need this.
-    resolveLoader: {
-      modules: [sourceNodeModules, 'node_modules']
+    node: {
+      dgram: 'empty',
+      fs: 'empty',
+      net: 'empty',
+      tls: 'empty',
+      child_process: 'empty'
     },
-    devtool: urc.production ? 'source-map' : 'cheap-module-source-map',
     optimization: {
       minimizer: [safeWebpackUglifier()],
       splitChunks: {
@@ -198,48 +79,32 @@ function createWebpackConfig(urc) {
       },
       runtimeChunk: 'single'
     },
-    node: {
-      dgram: 'empty',
-      fs: 'empty',
-      net: 'empty',
-      tls: 'empty',
-      child_process: 'empty'
+    output: {
+      path: path.join(urc.outputDirectory),
+      publicPath: path.join(urc.siteBasePath, '/'), // TOFIX clean up this patching (/) of siteBasePath and also publicAssetsApth
+      filename: urc.production
+        ? `${publicAssetsPath}/js/[name]-[chunkhash:10].js`
+        : `${publicAssetsPath}/js/[name].js`,
+      chunkFilename: urc.production
+        ? `${publicAssetsPath}/js/[name]-[chunkhash:10].chunk.js`
+        : `${publicAssetsPath}/js/[name].chunk.js`
     },
-    bail: urc.production,
-    performance: false
-  };
-  return urc.webpackConfigTransform(config);
-}
-
-function createBabelLoader(urc) {
-  const babelrcPath = path.join(urc.rootDirectory, '.babelrc');
-  const exists = fs.existsSync(babelrcPath);
-  const loader = {
-    loader: require.resolve('babel-loader'),
-    options: {
-      presets: [require('../../packages/babel-preset-mapbox')],
-      babelrc: false,
-      compact: false,
-      cacheDirectory: true,
-      cacheCompression: urc.production,
-      // `babel-preset-mapbox` depends on the `process.env.BROWSERSLIST` and
-      // any change in its value would fail to change the default `cacheIdentifier` of `babel-loader`
-      // hence leading to stale babel output. To mitigate this we need to create a more accurate `cacheIdentifier`
-      // which is (`defaultCacheIdentifier` + `urc.browserslist`) ref: https://github.com/babel/babel-loader/blob/7.x/src/index.js#L129
-      cacheIdentifier: JSON.stringify({
-        'babel-loader': require('babel-loader/package.json').version,
-        'babel-core': require('babel-core/package.json').version,
-        babelrc: exists ? fs.readFileSync(babelrcPath, 'utf8') : '',
-        env: process.env.BABEL_ENV || process.env.NODE_ENV || 'development',
-        browserslist: urc.browserslist
-      })
+    performance: false,
+    plugins: webpackPlugins(urc),
+    resolve: {
+      alias: {
+        // This is added to make Underreact symlinking work with npm <=3,
+        '@babel/runtime': path.dirname(
+          require.resolve('@babel/runtime/package.json')
+        )
+      }
+    },
+    // Loader names need to be strings, and to allow them to be looked
+    // up within this module's nested dependencies, not just the user's
+    // node_modules, we need this.
+    resolveLoader: {
+      modules: [sourceNodeModules, 'node_modules']
     }
   };
-
-  if (exists) {
-    delete loader.options.presets;
-    loader.options.babelrc = true;
-  }
-
-  return loader;
-}
+  return urc.webpackConfigTransform(config);
+};
